Fixed up dependency injection. The Web API is now functional again.
