+++ Skylight/Skylight-API/Models/Domain/BaseModel.cs
@@ -3,12 +3,10 @@
 namespace Skylight.Web.Models;
 
 /// <summary>
-/// Represents the shared properties of all view models.
+/// Represents the shared properties of all Web models.
 /// </summary>
 public abstract record BaseModel
 {
-    public int? Id { get; init; }
-
     [Required]
-    public required bool Deleted { get; init; }
+    public required Guid Id { get; init; }
 }
